requirements.txt
altair==4.2.2
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
Flask==2.3.1
Flask-RESTful==0.3.9
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
numpy==1.24.3
packaging==23.1
pandas==2.0.1
patsy==0.5.3
Pillow==9.5.0
plotly==5.14.1
plotly-express==0.4.1
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
rich==13.3.5
scipy==1.10.1
six==1.16.0
smmap==5.0.0
statsmodels==0.13.5
streamlit==1.22.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.3.2
zipp==3.15.0
